refactor: split CLI root surface into frontend module
